requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
arxiv>=2.1.0
pytest>=7.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
requests>=2.31.0
PyMuPDF>=1.24.0
